Train And Validation Datasets
This commit adds the new train and validation datasets! We were running into the limitations of tinyshakespear so time to move a dataset more representative of the original GPT2 and GPT3 datasets. Fineweb is a great candidate for this. Here we download it the 10B version of it i.e 10B tokens. Create the data shards for those tokens. Then we modify our training script to utilize these fineweb data shards, add a validation part, generate random samples part. We also need a standard eval, and for this we select hellasawg. we do the same for hellaswag and add in the hellaswag eval part. Lastly we add logging to record the results. And thats it! Time to run it now!